beamerthemeUppsala attribution
(This message belongs to only ‘gshs_beamer’.)
To whom it may concern:
Apologies for forgetting about attribution on original. We have adapted
Unofficial Uppsala theme about 1 year ago for our school students and
teachers, but at first we didn’t expected it to be open-sourced at
GitHub like this. Sorry for that.
By the way, you haven’t specified any license for the theme, so we
added LPPL-1.3c to this.